feat: :sparkles: se corrigen las migraciones

+++ database/migrations/2026_03_06_191335_create_carrito_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('carrito', function (Blueprint $table) {
             $table->integer('id', true);
-            $table->integer('usuario_id')->nullable()->index('idx_carrito_usuario');
+            $table->unsignedBigInteger('usuario_id')->nullable()->index('idx_carrito_usuario');
             $table->integer('producto_id')->nullable()->index('fk_carrito_producto');
             $table->integer('cantidad')->nullable()->default(1);
             $table->timestamp('fecha')->nullable()->useCurrent();
